\name{kopula}
\alias{kopula}

\title{
Copula object
}
\description{
Generate the copula object.
}
\usage{
kopula(fam, tht, dm)
}
\arguments{
  \item{fam}{
Family name copula. These can be: "clayton", "gumbel", "frank",
  "joe", "amh", "fgm".
}
  \item{tht}{
Copula parameter.
}
  \item{dm}{
Copula dimension.
}
}
\value{
Copula object
}
\author{
Josef Brejcha
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
