\name{prunikus}
\alias{prunikus}
\title{
The coordinates of the intersection lines of the cumulative distribution function and survival function
}
\description{
The coordinates of the intersection lines of the cumulative
   distribution function and survival function.
}
\usage{
prunikus(x, y)
}
\arguments{
  \item{x}{
Numeric vector of size 4. The horizontal coordinates of opposite points.
}
  \item{y}{
Numeric vector of size 4. The vertical coordinates of opposite points.
}
}

\value{
Numeric vector size 2.
}
\references{
Line-line intersection, \url{https://en.wikipedia.org/wiki/Line-line_intersection}
}
\author{
Josef Brejcha
}

\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
