\name{vfcp-package}
\alias{vfcp-package}
\docType{package}
\title{
Computation of v Values for U and Copula C(U, v)
}
\description{
Computation \code{v} when \code{u} and \code{C(u, v)} copula are
    known. Calculation and plotting of cumulative distribution
    and survival function when \code{u}, \code{C(u, v)} copula
    and marginal distributions are known. These calculations can
    be tabulated as option. The numerical definition of a common
    area limited by lines of the cumulative distribution function
    and survival function. Approximate quantification of the
    probability of this area. In addition to 'amh', the copula
    dimension may be larger than 2.
}
\details{
\tabular{ll}{
Package: \tab vfcp\cr
Type: \tab Package\cr
Version: \tab 1.4.0\cr
Date: \tab 2017-10-24\cr
License: \tab GPL (>= 3)\cr
}
}
\author{
Josef Brejcha

Maintainer: Josef Brejcha <brchjo@gmail.com>
}
\references{
A.K. SUZUKI, F. LOUZADA and V.G. CANCHO, On estimation and
influence diagnostics for a Bivariate Promotion Lifetime Model
Based on the FGM Copula: A Fully Bayesian Computation, \emph{Tendencias em Matematica Aplicada e Computacional, 14, N. 3 (2013), 441-461}, \url{http://www.scielo.br/pdf/tema/v14n3/a14v14n3.pdf} \cr \cr
M. Mahfoud, "Bivariate Archimedean copulas: an application to two stock market
indices", \emph{Vrije Universiteit Amsterdam}, \bold{BMI Paper}, Amsterdam-2012,
\url{http://docplayer.net/24882927-Bivariate-archimedean-copulas-an-application-to-two-stock-market-indices.html} \cr \cr
Copula (probability theory), \url{https://en.wikipedia.org/wiki/Copula_(probability_theory)}
Statistical - Distributions - Inverted Beta distribution - Example, \url{http://www.xycoon.com/ibeta.htm}
}
